Moved SegmentExpression into namespace Piwik\Segment
