+++ plugins/VisitFrequency/API.php
@@ -13,7 +13,7 @@ use Piwik\Archive;
 use Piwik\DataTable;
 use Piwik\Piwik;
 use Piwik\Plugins\VisitsSummary\API as APIVisitsSummary;
-use Piwik\SegmentExpression;
+use Piwik\Segment\SegmentExpression;
 
 /**
  * VisitFrequency API lets you access a list of metrics related to Returning Visitors.
